Fix: can not migrate

+++ database/migrations/2014_10_12_200000_add_two_factor_columns_to_users_table.php
@@ -16,17 +16,17 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             $table->text('two_factor_secret')
-                    ->after('password')
-                    ->nullable();
+                ->after('password')
+                ->nullable();
 
             $table->text('two_factor_recovery_codes')
-                    ->after('two_factor_secret')
-                    ->nullable();
+                ->after('two_factor_secret')
+                ->nullable();
 
             if (Fortify::confirmsTwoFactorAuthentication()) {
                 $table->timestamp('two_factor_confirmed_at')
-                        ->after('two_factor_recovery_codes')
-                        ->nullable();
+                    ->after('two_factor_recovery_codes')
+                    ->nullable();
             }
         });
     }
@@ -38,13 +38,15 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn(array_merge([
-                'two_factor_secret',
-                'two_factor_recovery_codes',
-            ], Fortify::confirmsTwoFactorAuthentication() ? [
-                'two_factor_confirmed_at',
-            ] : []));
-        });
+        if (Schema::hasColumn('users', 'two_factor_secret') && Schema::hasColumn('users', 'two_factor_recovery_codes')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn(array_merge([
+                    'two_factor_secret',
+                    'two_factor_recovery_codes',
+                ], Fortify::confirmsTwoFactorAuthentication() && Schema::hasColumn('users', 'two_factor_confirmed_at') ? [
+                    'two_factor_confirmed_at',
+                ] : []));
+            });
+        }
     }
 };
